+++ tests/cli/test_main.py
@@ -52,7 +52,6 @@ class TestCLIRun:
         assert command.env["WINEPREFIX"] == str(
             steam_dir / "steamapps" / "compatdata" / "4149337689" / "pfx")
 
-
     def test_run_winetricks_select_proton(
             self, cli, steam_app_factory, default_proton,
             custom_proton_factory, command, home_dir):
@@ -382,7 +381,10 @@ class TestCLIRun:
         assert len(caplog.records) == 0
 
         # Fake a Flatpak environment
-        monkeypatch.setattr("protontricks.cli.is_flatpak_sandbox", lambda: True)
+        monkeypatch.setattr(
+            "protontricks.cli.main.is_flatpak_sandbox",
+            lambda: True
+        )
 
         cli(["--no-bwrap", "-s", "nothing"])